Реализован главный servlet - QuestServlet и всё для его корректно функционирования

+++ src/main/java/com/javarush/quest/repository/GameRepository.java
@@ -5,5 +5,5 @@ import com.javarush.quest.entity.Question;
 import org.jetbrains.annotations.NotNull;
 
 public interface GameRepository {
-    Question read(@NotNull GameType type, int id);
+    Question read(@NotNull GameType type, long id);
 }
